moving gstpufs dir to gsmemcachedfs
