+++ tensorflow_io/core/kernels/gsmemcachedfs/BUILD
