+++ tensorflow_io/core/kernels/gsmemcachedfs/LICENSE
