+++ tensorflow_io/core/kernels/gsmemcachedfs/gs_tpu_file_system.h
@@ -21,8 +21,8 @@ limitations under the License.
 #include <vector>
 
 #include "absl/memory/memory.h"
-#include "tensorflow_io/core/kernels/gstpufs/gce_memcached_server_list_provider.h"
-#include "tensorflow_io/core/kernels/gstpufs/memcached_file_system.h"
+#include "tensorflow_io/core/kernels/gsmemcachedfs/gce_memcached_server_list_provider.h"
+#include "tensorflow_io/core/kernels/gsmemcachedfs/memcached_file_system.h"
 
 namespace tensorflow {
