+++ tensorflow_io/core/kernels/gsmemcachedfs/memcached_file_system.h
@@ -22,8 +22,8 @@ limitations under the License.
 
 #include "absl/memory/memory.h"
 #include "tensorflow/core/platform/cloud/gcs_file_system.h"
-#include "tensorflow_io/core/kernels/gstpufs/gce_memcached_server_list_provider.h"
-#include "tensorflow_io/core/kernels/gstpufs/memcached_file_block_cache.h"
+#include "tensorflow_io/core/kernels/gsmemcachedfs/gce_memcached_server_list_provider.h"
+#include "tensorflow_io/core/kernels/gsmemcachedfs/memcached_file_block_cache.h"
 
 namespace tensorflow {
